POSIX.xs: Fix to work better multi-threaded
This defines POSIX_EXT at the beginning of the file which automatically
causes any reentrant versions of the functions called here to be used.

It also adds critical sections around several function calls that need
them in multi-threaded applications.  These functions are all
locale-related, and while always broken, it was forbidden to use locales
in multi-threaded contexts, so the issue never came up.  This is in
preparation for making locales multi-thread capable.