Basic implementation of script runs
The hard part is getting the syntactic sugar to work so this is eacy to
use.

In the meantime, you have to do a script run by using \M which is a
zero-length assertion that the most recent previous capturing group is a
script run.

    qr/abc(def)\Mghi/

\M is a temporary arbitrary mechanism. used because it was available and
having such an assertion avoids the implementation problems the final
syntax has;  it would  probably be something more like

    qr/abc(?*SCRIPT_RUN:def)ghi/